export collectionClass from 'Collection' module
